[폴더위계정리] refactor: SNUTTStringUtils + SNUTTStringUtilsNew 통합
두 object를 해체하고 역할별 top-level 함수 파일로 분리:
- LectureFormatterExt.kt: 강의 표시 텍스트 (시간, 위치, 태그 등)
- TimeFormatterExt.kt: 시간/날짜 포매팅
- ValidationExt.kt: 이메일/비밀번호/ID 검증
