+++ app/src/main/java/com/wafflestudio/snutt2/data/mapper/NotificationMappers.kt
@@ -3,12 +3,12 @@ package com.wafflestudio.snutt2.data.mapper
 import com.wafflestudio.snutt2.domain.model.Notification
 import com.wafflestudio.snutt2.domain.model.NotificationType
 import com.wafflestudio.snutt2.network.dto.NotificationDto
-import com.wafflestudio.snutt2.ui.util.SNUTTStringUtils
+import com.wafflestudio.snutt2.ui.util.getLocalDateTimeFromString
 
 fun NotificationDto.toDomain(): Notification = Notification(
     title = title,
     message = message,
-    createdAt = SNUTTStringUtils.getLocalDateTimeFromString(createdAt),
+    createdAt = getLocalDateTimeFromString(createdAt),
     type = when (type) {
         0 -> NotificationType.Warning
         1 -> NotificationType.Calendar
