+++ app/src/main/java/com/wafflestudio/snutt2/feature/login/FindIdPage.kt
@@ -33,7 +33,7 @@ import com.wafflestudio.snutt2.ui.components.compose.WebViewStyleButton
 import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
-import com.wafflestudio.snutt2.ui.util.SNUTTStringUtils.isEmailInvalid
+import com.wafflestudio.snutt2.ui.util.isEmailInvalid
 import com.wafflestudio.snutt2.ui.util.toast
 
 @Composable
