+++ app/src/main/java/com/wafflestudio/snutt2/feature/login/SignUpPage.kt
@@ -49,8 +49,8 @@ import com.wafflestudio.snutt2.ui.components.compose.WebViewStyleButton
 import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
-import com.wafflestudio.snutt2.ui.util.SNUTTStringUtils.isIdInvalid
-import com.wafflestudio.snutt2.ui.util.SNUTTStringUtils.isPasswordInvalid
+import com.wafflestudio.snutt2.ui.util.isIdInvalid
+import com.wafflestudio.snutt2.ui.util.isPasswordInvalid
 import com.wafflestudio.snutt2.ui.util.toast
 
 @Composable
