+++ app/src/main/java/com/wafflestudio/snutt2/feature/login/reset_password/EnterFullEmailStep.kt
@@ -34,7 +34,7 @@ import com.wafflestudio.snutt2.ui.components.compose.WebViewStyleButton
 import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
-import com.wafflestudio.snutt2.ui.util.SNUTTStringUtils.isEmailInvalid
+import com.wafflestudio.snutt2.ui.util.isEmailInvalid
 import com.wafflestudio.snutt2.ui.util.toast
 
 @Composable
