+++ app/src/main/java/com/wafflestudio/snutt2/feature/notifications/NotificationScreen.kt
@@ -48,7 +48,7 @@ import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
 import com.wafflestudio.snutt2.ui.theme.isDarkMode
-import com.wafflestudio.snutt2.ui.util.SNUTTStringUtils.getNotificationTime
+import com.wafflestudio.snutt2.ui.util.getNotificationTime
 import kotlinx.coroutines.flow.flowOf
 
 @Composable
